move to double precision for most things

+++ src/cedar_sky_trait.rs
@@ -11,14 +11,14 @@ pub trait CedarSkyTrait {
     fn get_object_types(&self) -> Vec<ObjectType>;
     fn get_constellations(&self) -> Vec<Constellation>;
     fn query_catalog_entries(&self,
-                             max_distance: Option<f32>,
-                             min_elevation: Option<f32>,
-                             faintest_magnitude: Option<f32>,
+                             max_distance: Option<f64>,
+                             min_elevation: Option<f64>,
+                             faintest_magnitude: Option<f64>,
                              catalog_match: &Vec<String>,
                              object_type_match: &Vec<String>,
                              ordering: Option<Ordering>,
-                             dedup_distance: Option<f32>,
-                             decrowd_distance: Option<f32>,
+                             dedup_distance: Option<f64>,
+                             decrowd_distance: Option<f64>,
                              sky_location: Option<&CelestialCoord>,
                              location_info: Option<&LocationInfo>)
                              -> Result<Vec<SelectedCatalogEntry>, CanonicalError>;
